mergeme: Don't overwrite unmodified files that have a matching md5sum
If a file has not not been modified (according to its mtime) and the
installed instance's md5sum matches the new md5sum then there is
little point in overwriting these files. Not doing so results in much
less disk churn, especially when PORTAGE_TMPDIR is on a different
filesystem. Think of the SSDs!